fix(db): align menu_beverage_options.menu_id with menus.id (INT)
Migration 113 declared menu_id BIGINT but menus.id is INT (signed), causing
MySQL error 3780 on startup: 'Referencing column menu_id and referenced
column id ... are incompatible.'

Change menu_id column type to INT to match menus.id (consistent with
migration 045 which also uses INT).

+++ internal/db/migrations/113_menu_beverage_options.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS restaurant_beverage_options (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 CREATE TABLE IF NOT EXISTS menu_beverage_options (
-  menu_id BIGINT NOT NULL,
+  menu_id INT NOT NULL,
   restaurant_id INT NOT NULL,
   beverage_option_id BIGINT NOT NULL,
   selected TINYINT(1) NOT NULL DEFAULT 1,
